feat: add show on map for the favourite countries as well
